#63 Cafe Comment count 정리

+++ src/main/java/com/widehouse/cafe/article/entity/Article.java
@@ -58,7 +58,7 @@ public class Article {
             inverseJoinColumns = @JoinColumn(name = "tag_id", referencedColumnName = "id"))
     private List<Tag> tags;
 
-    private int commentCount;
+    private long commentCount;
 
     @CreatedDate
     private LocalDateTime createdAt;
